Phase 4 batch 4: extract AbandonControl and PlayCard from SubmissionStatusView
Splits the abandon-confirmation control and the playable-build CTA
card out of SubmissionStatusView.tsx, which drops from 1,264 to 1,161
lines. JSDoc blocks on both components are rewritten as compliant
single-line comments.
